Refactor GreatswordItem from aura to regular item
Move GreatswordItem out of the aura package, remove the
AuraItemInterface and its particle aura effect, and update the recipe
import accordingly. The swing-based soul
